+++ lib/recipes/data-science/recipe.rb
@@ -135,6 +135,7 @@ do |doc, _resources|
     BakeAppendix.v1(page: page, number: appendix_letter)
   end
 
+  BakeTableColumns.v1(book: book)
   BakeEquations.v1(book: book)
   BakeIndex.v1(book: book)
   BakeFootnotes.v1(book: book)